Refactor logger and background script structure
- Removed the old logger implementation and added a new logger in the infrastructure directory.
- Updated imports in main and react-inject files to use the new logger path.
- Deleted outdated test files for background and calendar manager.
- Created a new background script to handle message communication with content scripts.
- Added tests for the new background script to ensure message listener functionality.
