--- index.html
@@ -1,28 +1,0 @@
-<html>
-    <head>
-        <script src="lib/jquery/jquery.min.js"></script>
-        <script src="lib/mdl/material.js"></script>
-        <script src="src/calendar_manager.js"></script>
-        <script src="lib/vue/vue.js"></script>
-        <script src="lib/vue/vuetify.js"></script>
-        <script src="lib/mousetrap/mousetrap.min.js"></script>
-        <script src="lib/mousetrap/mousetrap-global-bind.min.js"></script>
-        <script src="src/inject/inject.js"></script>
-
-        <link rel="stylesheet" href="lib/mdl/material.css">
-        <link rel="stylesheet" href="src/inject/inject.css">
-
-        <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700|Material+Icons">
-        <link rel="stylesheet" href="lib/vue/vuetify_scoped.css">
-
-
-    </head>
-    <body style="background: rgb(255, 200, 200)">
-        <div id="calendar_selector_ui">
-        </div>
-    </body>
-
-    <script>
-      insertUI($('#calendar_selector_ui'))
-    </script>
-</html>
