+++ manifest.json
@@ -13,7 +13,7 @@
     "128": "icons/icon128.png"
   },
   "background": {
-    "service_worker": "src/background.ts"
+    "service_worker": "src/infrastructure/background.ts"
   },
   "web_accessible_resources": [
     {
