--- scratchpad/vue-test.html
@@ -1,138 +1,0 @@
-<html>
-  <head>
-    <!-- <script src="https://unpkg.com/vue"></script> -->
-
-    <link href='https://fonts.googleapis.com/css?family=Roboto:300,400,500,700|Material+Icons' rel="stylesheet">
-    <link href="https://unpkg.com/vuetify/dist/vuetify.min.css" rel="stylesheet">
-  </head>
-  
-  <body>
-    <v-app id="app">
-      <main>
-        <!-- <v-container fluid grid-list-md text-xs-center> -->
-          <v-layout row>
-            <v-flex md6 class="pt-3">
-              <span class="btn-toggle">
-                <template  v-for="button in buttons">
-                  <v-tooltip bottom open-delay="1000" transition="false">
-                    <v-btn small slot="activator" @click="btn_clicked(button)">{{ button.text }}</v-btn>
-                    <span>{{ button.tooltip }}</span>
-                  </v-tooltip>
-                </template>
-
-                <v-menu bottom 
-                        offset-y
-                        :close-on-content-click="false"
-                        v-model="presets_menu_open"
-                        ref="presets_menu"
-                >
-                  <v-btn small slot="activator" @click="presets_open()"> Presets</v-btn>
-                  <v-select
-                    class="select"
-                    v-bind:items="dropdown"
-                    label="Load preset..."
-                    editable
-                    small
-                    return-object
-                    hide-details
-                    item-value="text"
-                    ref="select"
-                    @input="select_input"
-                  >
-                    <template slot="item" slot-scope="data">
-                      <v-list-tile-content>
-                        <v-list-tile-title v-html="data.item.text"></v-list-tile-title>
-                      </v-list-tile-content>
-
-                      <v-list-tile-action>
-                        <v-btn icon ripple @click.stop="select_delete(data.item)">
-                          <v-icon color="grey lighten-1">delete</v-icon>
-                        </v-btn>
-                      </v-list-tile-action>
-                    </template>
-                  </v-select>
-
-                </v-menu>
-              </span>
-            </v-flex>
-          </v-layout>
-          <!-- </v-container> -->
-      </main>
-    </v-app>
-  </body>
-  
-  <script src="https://vuejs.org/js/vue.js"></script>
-  <script src="https://unpkg.com/vuetify/dist/vuetify.js"></script>
-  <!-- <script src="https://code.jquery.com/jquery-3.2.1.min.js"></script> -->
-               <!-- <script src="vue-test.js"></script> -->
-
-  <script>
-   var my_list = [
-     'Learn JavaScript',
-     'Learn Vue',
-     'Build something awesome'
-   ]
-
-   var app = new Vue({
-     el: '#app',
-     data: {
-       presets_menu_open: false,
-       my_list: my_list,
-       buttons: [
-         {text: 'User', tooltip: 'Enable a user by name or regexp'},
-         {text: "Save As", tooltip: 'Save current calendars as a named preset' },
-         {text: "Restore", tooltip: 'Restore previous calendars (set by Load & Clear)' },
-         {text: "Clear", tooltip: 'Clear all calendars' },
-       ]
-     },
-     methods: {
-       presets_open: function(){
-         setTimeout(()=>{ // timeout to allow the menu to be rendered first
-           /* console.log(this.$refs.select)*/
-           if(this.$refs.presets_menu.isActive){
-             this.$refs.select.focusInput()
-             this.$refs.select.showMenu()
-           }
-         }, 100)
-       },
-       select_input: function(value) {
-         console.log('input', value && value.text, value)
-         console.log(this.$refs.presets_menu)
-         console.log(this)
-         this.presets_menu_open = false
-       },
-       select_delete: function(item){
-         console.log('DELETE', item.text, item)
-         const del = confirm('Are you sure you want to delete this preset?')
-         if(del){
-           setTimeout(() => {
-             const index = my_list.indexOf(item.text)
-             if(index >= 0){
-               my_list.splice(index, 1)
-             }
-           }, 800)
-         }
-       },
-       btn_clicked: function(button){
-         console.log('clicked on', button.text, button)
-       },
-     },
-     computed: {
-       dropdown: function () {
-         return this.my_list.map( (e) => { 
-           return {
-             text: e,
-             subtitle: `sub for ${e} that is really, really, really long`,
-           } 
-         } )
-       }
-     }
-   })
-
-
-   setTimeout(()=>{
-     my_list.push('another entry')
-   }, 5000)
-
-  </script>
-</html>
