--- tests/calendar_manager.test.ts
@@ -1,338 +1,0 @@
-import { describe, it, expect, jest, beforeEach } from '@jest/globals';
-
-describe('calendar_manager.ts', () => {
-  // Mock DOM setup
-  beforeEach(() => {
-    // Reset document body
-    document.body.innerHTML = '';
-    
-    // Mock jQuery
-    (global as any).jQuery = jest.fn((selector: any) => {
-      if (typeof selector === 'string') {
-        // Return a mock jQuery object
-        const mockJQuery = {
-          width: jest.fn().mockReturnThis(),
-          height: jest.fn().mockReturnThis(),
-          prependTo: jest.fn().mockReturnThis(),
-          addClass: jest.fn().mockReturnThis(),
-          css: jest.fn().mockReturnThis(),
-          show: jest.fn().mockReturnThis(),
-          hide: jest.fn().mockReturnThis(),
-          animate: jest.fn((_props: any, options: any) => {
-            if (options && options.complete) {
-              setTimeout(options.complete, 0);
-            }
-            return mockJQuery;
-          }),
-        };
-        return mockJQuery;
-      }
-      // If passed an element, return a mock jQuery object wrapping it
-      return {
-        width: () => 100,
-        height: () => 100,
-        animate: jest.fn((_props: any, options: any) => {
-          if (options && options.complete) {
-            setTimeout(options.complete, 0);
-          }
-        }),
-      };
-    });
-  });
-
-  describe('Utility Functions', () => {
-    it('should export sleep function that waits specified milliseconds', async () => {
-      // Import the module to test utility functions
-      // Since we can't directly access the IIFE, we'll test through CalendarManager
-      
-      // This is a placeholder test since sleep is internal
-      expect(true).toBe(true);
-    });
-  });
-
-  describe('Overlay Class', () => {
-    it('should be accessible through CalendarManager', () => {
-      // Since the code runs in an IIFE, we can't directly test the classes
-      // We would need to refactor the code to export classes for unit testing
-      // For now, we'll create integration-style tests
-      expect(true).toBe(true);
-    });
-  });
-
-  describe('CalendarDOM Class', () => {
-    it('should check if element is attached to DOM', () => {
-      // Create a mock calendar list item
-      const li = document.createElement('li');
-      const dataDiv = document.createElement('div');
-      dataDiv.setAttribute('data-id', btoa('test@example.com'));
-      const checkbox = document.createElement('input');
-      checkbox.type = 'checkbox';
-      checkbox.setAttribute('aria-label', 'Test Calendar');
-      checkbox.checked = true;
-      
-      li.appendChild(dataDiv);
-      li.appendChild(checkbox);
-      
-      // Test that the element is not attached initially
-      expect(document.body.contains(li)).toBe(false);
-      
-      // Attach to document
-      document.body.appendChild(li);
-      expect(document.body.contains(li)).toBe(true);
-    });
-  });
-
-  describe('Calendar Class', () => {
-    it('should parse calendar data from DOM element', () => {
-      // Create a mock calendar list item
-      const li = document.createElement('li');
-      const dataDiv = document.createElement('div');
-      const email = 'test@example.com';
-      dataDiv.setAttribute('data-id', btoa(email));
-      const checkbox = document.createElement('input');
-      checkbox.type = 'checkbox';
-      const calName = 'Test Calendar';
-      checkbox.setAttribute('aria-label', calName);
-      checkbox.checked = true;
-      
-      li.appendChild(dataDiv);
-      li.appendChild(checkbox);
-      document.body.appendChild(li);
-      
-      // The Calendar constructor would parse this
-      expect(atob(dataDiv.getAttribute('data-id')!)).toBe(email);
-      expect(checkbox.getAttribute('aria-label')).toBe(calName);
-      expect(checkbox.checked).toBe(true);
-    });
-  });
-
-  describe('CalendarList Class', () => {
-    it('should extend Array', () => {
-      const list: any[] = [];
-      expect(Array.isArray(list)).toBe(true);
-      expect(list).toBeInstanceOf(Array);
-    });
-
-    it('should support filter operations', () => {
-      const items = [
-        { id: '1', name: 'Cal 1', checked: true },
-        { id: '2', name: 'Cal 2', checked: false },
-        { id: '3', name: 'Cal 3', checked: true },
-      ];
-      
-      const enabled = items.filter(item => item.checked);
-      const disabled = items.filter(item => !item.checked);
-      
-      expect(enabled.length).toBe(2);
-      expect(disabled.length).toBe(1);
-    });
-
-    it('should support map operations', () => {
-      const items = [
-        { id: '1', name: 'Cal 1' },
-        { id: '2', name: 'Cal 2' },
-      ];
-      
-      const ids = items.map(item => item.id);
-      expect(ids).toEqual(['1', '2']);
-    });
-  });
-
-  describe('CalendarManager', () => {
-    it('should have groups object', () => {
-      const groups: { [key: string]: string[] } = {};
-      groups['work'] = ['cal1@example.com', 'cal2@example.com'];
-      groups['personal'] = ['personal@example.com'];
-      
-      expect(groups['work']).toHaveLength(2);
-      expect(groups['personal']).toHaveLength(1);
-    });
-
-    it('should support regex pattern for excluding groups', () => {
-      const excludeRe = /^(saved_|__)/;
-      
-      expect('saved_123'.match(excludeRe)).toBeTruthy();
-      expect('__last_saved'.match(excludeRe)).toBeTruthy();
-      expect('work'.match(excludeRe)).toBeFalsy();
-      expect('personal'.match(excludeRe)).toBeFalsy();
-    });
-
-    it('should filter internal groups when exporting', () => {
-      const groups: { [key: string]: string[] } = {
-        'work': ['cal1@example.com'],
-        'personal': ['personal@example.com'],
-        'saved_123': ['cal1@example.com'],
-        '__last_saved': ['work'],
-      };
-      
-      const excludeRe = /^(saved_|__)/;
-      const exportedGroups = { ...groups };
-      
-      Object.keys(exportedGroups)
-        .filter(name => name.match(excludeRe))
-        .forEach(name => delete exportedGroups[name]);
-      
-      expect(Object.keys(exportedGroups)).toEqual(['work', 'personal']);
-      expect(exportedGroups['saved_123']).toBeUndefined();
-      expect(exportedGroups['__last_saved']).toBeUndefined();
-    });
-
-    it('should generate timestamp-based autosave names', () => {
-      const now = Date.now();
-      const groupName = `saved_${now}`;
-      
-      expect(groupName).toMatch(/^saved_\d+$/);
-    });
-
-    it('should maintain last saved array', () => {
-      const lastSaved: string[] = [];
-      
-      lastSaved.push('work');
-      lastSaved.push('personal');
-      lastSaved.push('saved_123');
-      
-      expect(lastSaved).toHaveLength(3);
-      
-      const restored = lastSaved.pop();
-      expect(restored).toBe('saved_123');
-      expect(lastSaved).toHaveLength(2);
-    });
-
-    it('should support filtering calendars by group', () => {
-      const groupIds = ['cal1@example.com', 'cal2@example.com'];
-      const calendars = [
-        { id: 'cal1@example.com', name: 'Calendar 1', checked: false },
-        { id: 'cal2@example.com', name: 'Calendar 2', checked: false },
-        { id: 'cal3@example.com', name: 'Calendar 3', checked: true },
-      ];
-      
-      const inGroup = calendars.filter(c => 
-        groupIds.indexOf(c.id) >= 0 || groupIds.indexOf(c.name) >= 0
-      );
-      
-      const notInGroup = calendars.filter(c => 
-        groupIds.indexOf(c.id) < 0 && groupIds.indexOf(c.name) < 0
-      );
-      
-      expect(inGroup).toHaveLength(2);
-      expect(notInGroup).toHaveLength(1);
-    });
-  });
-
-  describe('Scroll Functions', () => {
-    it('should calculate scroll position delta', () => {
-      const delta = 0.000002;
-      const eq = (a: number, b: number) => Math.abs(a - b) < delta;
-      
-      expect(eq(0, 0)).toBe(true);
-      expect(eq(1, 1.000001)).toBe(true);
-      expect(eq(1, 2)).toBe(false);
-      expect(eq(100, 100.000001)).toBe(true);
-    });
-
-    it('should calculate scroll increments', () => {
-      const scrollTop = 0;
-      const scrollIncrement = 50;
-      
-      const positions = [];
-      let current = scrollTop;
-      
-      for (let i = 0; i < 5; i++) {
-        current += scrollIncrement;
-        positions.push(current);
-      }
-      
-      expect(positions).toEqual([50, 100, 150, 200, 250]);
-    });
-  });
-
-  describe('Data Encoding', () => {
-    it('should encode and decode calendar IDs', () => {
-      const email = 'test@example.com';
-      const encoded = btoa(email);
-      const decoded = atob(encoded);
-      
-      expect(decoded).toBe(email);
-    });
-
-    it('should handle various email formats', () => {
-      const emails = [
-        'user@example.com',
-        'user.name@example.co.uk',
-        'user+label@example.com',
-        'user_name@sub.example.com',
-      ];
-      
-      emails.forEach(email => {
-        const encoded = btoa(email);
-        const decoded = atob(encoded);
-        expect(decoded).toBe(email);
-      });
-    });
-  });
-
-  describe('RegExp Patterns', () => {
-    it('should match calendar names with regex', () => {
-      const calendars = [
-        { name: 'Work Calendar' },
-        { name: 'Personal Calendar' },
-        { name: 'Team Meetings' },
-        { name: 'Holidays' },
-      ];
-      
-      const workRegex = new RegExp('work', 'i');
-      const matched = calendars.filter(c => c.name.match(workRegex));
-      
-      expect(matched).toHaveLength(1);
-      expect(matched[0].name).toBe('Work Calendar');
-    });
-
-    it('should match multiple calendars with broader pattern', () => {
-      const calendars = [
-        { name: 'Work Calendar' },
-        { name: 'Personal Calendar' },
-        { name: 'Team Meetings' },
-      ];
-      
-      const calRegex = new RegExp('calendar', 'i');
-      const matched = calendars.filter(c => c.name.match(calRegex));
-      
-      expect(matched).toHaveLength(2);
-    });
-
-    it('should match all calendars with dot pattern', () => {
-      const calendars = [
-        { name: 'Work' },
-        { name: 'Personal' },
-        { name: 'Team' },
-      ];
-      
-      const allRegex = new RegExp('.', 'i');
-      const matched = calendars.filter(c => c.name.match(allRegex));
-      
-      expect(matched).toHaveLength(3);
-    });
-  });
-
-  describe('Storage Format', () => {
-    it('should structure groups correctly', () => {
-      const storage: { [key: string]: any } = {};
-      
-      // Add regular groups
-      storage['work'] = ['cal1@example.com', 'cal2@example.com'];
-      storage['personal'] = ['personal@example.com'];
-      
-      // Add autosave
-      storage['saved_1234567890'] = ['cal1@example.com'];
-      
-      // Add internal tracking
-      storage['__last_saved'] = ['work', 'personal'];
-      storage['__v'] = 1;
-      
-      expect(Object.keys(storage)).toContain('work');
-      expect(Object.keys(storage)).toContain('__last_saved');
-      expect(Array.isArray(storage['__last_saved'])).toBe(true);
-      expect(typeof storage['__v']).toBe('number');
-    });
-  });
-});
